Refactor import so functions can be defined in ncu.py
