+++ thicket/helpers.py
@@ -3,9 +3,6 @@
 #
 # SPDX-License-Identifier: MIT
 
-import re
-from difflib import SequenceMatcher
-
 from more_itertools import powerset
 import pandas as pd
 
@@ -171,111 +168,3 @@ def _get_perf_columns(df):
 def _powerset_from_tuple(tup):
     pset = [y for y in powerset(tup)]
     return {x[0] if len(x) == 1 else x for x in pset}
-
-
-def _match_call_trace_regex(
-    kernel_call_trace, demangled_kernel_name, debug, action=None
-):
-    """Use the NCU call trace to regex match the kernel name from the demangled
-    kernel string. Also modifies the demangled kernel name in certain cases. Returns
-    the matched kernel string, if match is possible.
-
-    Arguments:
-        kernel_call_trace (list): List of strings from NCU representing the call trace
-        demangled_kernel_name (str): Demangled kernel name from NCU
-        debug (bool): Print debug statements
-        action (ncu_report.IAction): NCU action object
-    """
-    # Call trace with last element removed
-    # (last elem usually not useful for matching)
-    temp_call_trace = kernel_call_trace[:-1]
-    # Special case to match "cub" kernels
-    if "cub" in demangled_kernel_name:
-        call_trace_str = "cub"
-        # Replace substrings that may cause mismatch
-        demangled_kernel_name = demangled_kernel_name.replace("(bool)1", "true")
-        demangled_kernel_name = demangled_kernel_name.replace("(bool)0", "false")
-    else:
-        call_trace_str = "::".join([s.lower() for s in temp_call_trace])
-    if debug:
-        print(f"\tKernel Call Trace: {kernel_call_trace}")
-        print(f"\t{action.name()}")
-
-    # Pattern ends with ":" if RAJA_CUDA, "<" if Base_CUDA
-    kernel_pattern = rf"{call_trace_str}::(\w+)[<:]"
-    kernel_match = re.search(kernel_pattern, demangled_kernel_name)
-    # Found match
-    if kernel_match:
-        kernel_str = kernel_match.group(1)
-    else:
-        if debug:
-            print(f"\tCould not match {demangled_kernel_name}")
-        return None, None, None, True
-
-    # RAJA_CUDA/Lambda_CUDA variant
-    instance_pattern = r"instance (\d+)"
-    instance_match = re.findall(instance_pattern, demangled_kernel_name)
-    if instance_match:
-        instance_num = instance_match[-1]
-        instance_exists = True
-    else:
-        # Base_CUDA variant
-        instance_num = None
-        instance_exists = False
-
-    return kernel_str, demangled_kernel_name, instance_num, instance_exists, False
-
-
-def _match_kernel_str_to_cali(
-    node_set, kernel_str, instance_num, raja_lambda_cuda, instance_exists
-):
-    """Given a set of nodes, node_set, from querying the Caliper call
-    tree using the NCU call trace, match the kernel_str to one of the
-    node names. Additionally, use the instance number, instance_num to
-    match kernels with multiple instances, if applicable.
-
-    Arguments:
-        node_set (list): List of Hatchet nodes from querying the call tree
-        kernel_str (str): Kernel name from _match_call_trace_regex
-        instance_num (int): Instance number of kernel, if applicable
-        raja_lambda_cuda (bool): True if RAJA_CUDA or Lambda_CUDA, False if Base_CUDA
-        instance_exists (bool): True if instance number exists, False if not
-    """
-    return [
-        n
-        for n in node_set
-        if kernel_str in n.frame["name"]
-        and (
-            f"#{instance_num}" in n.frame["name"]
-            if raja_lambda_cuda and instance_exists
-            else True
-        )
-    ]
-
-
-def _multi_match_fallback_similarity(matched_nodes, demangled_kernel_name, debug):
-    """If _match_kernel_str_to_cali has more than one match, attempt to match using sequence similarity.
-
-    Arguments:
-        matched_nodes (list): List of matched Hatchet nodes
-        demangled_kernel_name (str): Demangled kernel name from _match_call_trace_regex
-        debug (bool): Print debug statements
-
-    Returns:
-        matched_node (Hatchet.node): Hatchet node with highest similarity score
-    """
-    # Attempt to match using similarity
-    match_dict = {}
-    for node in matched_nodes:
-        match_ratio = SequenceMatcher(
-            None, node.frame["name"], demangled_kernel_name
-        ).ratio()
-        match_dict[match_ratio] = node
-    # Get highest ratio
-    highest_ratio = max(list(match_dict.keys()))
-    matched_node = match_dict[highest_ratio]
-    if debug:
-        print(
-            f"NOTICE: Multiple matches ({len(matched_nodes)}) found for kernel. Matching using string similarity..."
-        )
-    return matched_node
